# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMSOne Enterprise API
  description: |-
    Enterprise Spring Modulith template API. Every response uses the unified envelope ({data | errors, meta, links}) with meta.requestId always present; quote the requestId when reporting issues.

    Operations are grouped as `<axis> · <resource>`. The axis is the authority the operation enforces — Platform (a realm role), Organization (a permission inside one tenant), Shared (any authenticated caller) — and it is read from the code that enforces it, so a group name cannot promise access the API will refuse.
  contact:
    name: SMSOne
    email: ayesigapo@gmail.com
  version: "1"
servers:
- url: http://localhost:8080
  description: Local
- url: https://staging-api.smsone.co.ug
  description: Staging
- url: https://api.smsone.co.ug
  description: Production
security:
- bearerAuth: []
- keycloak: []
...
paths:
...
  /api/v1/orgs/{orgId}/members/{subject}/role:
    put:
      tags:
      - Organization · Members
      summary: Reassign a member's role
      description: "Last-owner protected: demoting the only remaining OWNER is a 409,\
        \ because it would lock the organization out of its own administration. The\
        \ caller must already hold every permission the new role carries, so this\
        \ endpoint alone cannot escalate. Assigning the role the member already has\
        \ is a no-op."
      operationId: assignRole
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: subject
        in: path
        required: true
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleRequest"
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
...
  /api/v1/orgs/{orgId}/members:
    get:
      tags:
      - Organization · Members
      summary: List organization members
      operationId: list_2
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: "page[size]"
        in: query
        description: "Items per page, 1–100. Defaults to 20."
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: "page[after]"
        in: query
        description: Opaque keyset cursor taken from a previous response's meta.page.nextCursor.
          Omit for the first page.
        required: false
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WindowedResultResourceObject"
    post:
      tags:
      - Organization · Members
      summary: Invite a member to the organization
      description: "One call does three things: it provisions the identity (creating\
        \ the Keycloak account and e-mailing temporary credentials if the person is\
        \ new), links them to the Keycloak organization, and records the membership.\
        \ Re-inviting an existing member is idempotent and returns the current membership\
        \ unchanged — it does not re-role them. Handing over a role is granting its\
        \ permissions, so the caller must already hold every permission `roleCode`\
        \ carries."
      operationId: invite
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
        required: true
      responses:
        "201":
          description: Created
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
...
  /api/v1/orgs/{orgId}/members/{subject}:
    delete:
      tags:
      - Organization · Members
      summary: Remove a member from the organization
      description: "Unlinks the membership and the Keycloak organization link only\
        \ — the person's Keycloak account is never deleted, so they keep their identity\
        \ and any membership of other organizations. Removing the last OWNER is a\
        \ 409."
      operationId: remove
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: subject
        in: path
        required: true
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
...
components:
  schemas:
...
    ResourceObject:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        attributes: {}
...
    AssignRoleRequest:
      type: object
      properties:
        roleCode:
          type: string
          minLength: 1
      required:
      - roleCode
...
    InviteMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        firstName:
          type: string
        lastName:
          type: string
        roleCode:
          type: string
          minLength: 1
      required:
      - email
      - roleCode
...
    PageMeta:
      type: object
      properties:
        size:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
        hasMore:
          type: boolean
        nextCursor:
          type: string
    WindowedResultResourceObject:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceObject"
        page:
          $ref: "#/components/schemas/PageMeta"
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste a Keycloak-issued access token.
      scheme: bearer
      bearerFormat: JWT
    keycloak:
      type: oauth2
      description: Run the Keycloak authorization-code flow (PKCE).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/auth
          tokenUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect
            profile: Profile claims